feat(api): switch to esm
